test: adapt tests to added functionalities

+++ app/src/main/java/com/github/onlynotesswent/ui/overview/editnote/EditNote.kt
@@ -106,7 +106,7 @@ fun EditNoteScreen(
 
   Scaffold(
       floatingActionButton = {
-        if (note!!.isOwner(currentUser!!.uid)) {
+        if (note != null && note!!.isOwner(currentUser!!.uid)) {
           DeleteButton(currentUser, note, navigationActions, noteViewModel)
         }
       },
